# KC_API/api/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "KickConnect API",
    "description": "Auto-generated Swagger documentation",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/get-durations": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/get-reservationCounts": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/get-main-schedule": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/get-location-assignment-schedule/{locationId}": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "description": "",
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/get-next-class/{locationId}": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "description": "",
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/get-location-class-schedule/{locationId}/{accountId}": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "description": "",
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/update-schedule/{scheduleMainId}": {
      "put": {
        "tags": [
          "Schedule"
        ],
        "description": "",
        "parameters": [
          {
            "name": "scheduleMainId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "selectedDate": {
                  "example": "any"
                },
                "startTime": {
                  "example": "any"
                },
                "endTime": {
                  "example": "any"
                },
                "day": {
                  "example": "any"
                },
                "locationValues": {
                  "example": "any"
                },
                "accountId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/add-schedule": {
      "post": {
        "tags": [
          "Schedule"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/delete-schedule-event/{scheduleMainId}": {
      "delete": {
        "tags": [
          "Schedule"
        ],
        "description": "",
        "parameters": [
          {
            "name": "scheduleMainId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  }
}